refactor:Update router registration and type hints in NexiosApp
Ensure proper router registration and add type hints for clarity in the NexiosApp class.
