+++ nexios/application.py
@@ -120,7 +120,7 @@ class NexiosApp(object):
         self.exceptions_handler = ExceptionMiddleware()
         self.server_error_handler = server_error_handler
 
-        self.app = Router(routes=routes)
+        self.app = Router(routes=routes) # type:ignore
         self.router = self.app
         self.route = self.router.route
         self.lifespan_context: Optional[lifespan_manager] = lifespan
@@ -302,7 +302,7 @@ class NexiosApp(object):
         """Set up automatic OpenAPI documentation"""
         docs = self.docs
         for route in self.get_all_routes():
-            if route.exlude_from_schema:
+            if getattr(self,"exlude_from_schema") == True:
                 continue
             for method in route.methods:
 
@@ -2055,5 +2055,11 @@ class NexiosApp(object):
             handler=handler,
         )
 
+    def register(
+        self,
+        app: ASGIApp,
+        prefix: str = "",
+    ) -> None:
+        self.router.register(app, prefix)
     def __str__(self) -> str:
         return f"<NexiosApp: {self.title}>"
