+++ nexios/routing/_utils.py
@@ -1,0 +1,18 @@
+from nexios.types import Scope
+
+def get_route_path(scope: Scope) -> str:
+    path: str = scope["path"]
+    root_path = scope.get("root_path", "")
+    if not root_path:
+        return path
+
+    if not path.startswith(root_path):
+        return path
+
+    if path == root_path:
+        return ""
+
+    if path[len(root_path)] == "/":
+        return path[len(root_path) :]
+
+    return path
