+++ nexios/routing/http.py
@@ -1,156 +1,31 @@
 from __future__ import annotations
+from typing import Any, List, Dict, Optional, Callable, Union, Pattern, Tuple, Annotated, Type
+from typing_extensions import Literal,Doc
 import inspect
-from typing import (
-    Any,
-    List,
-    Optional,
-    Pattern,
-    Dict,
-    TypeVar,
-    Tuple,
-    Callable,
-    Union,
-    Type,
-)
-from dataclasses import dataclass
-import re
 import copy
-import warnings, typing
-from enum import Enum
-from abc import abstractmethod, ABC
-import asyncio
-from nexios.events import AsyncEventEmitter
-from nexios.openapi.models import Parameter
-from nexios.types import MiddlewareType, WsMiddlewareType, HandlerType, WsHandlerType
-from nexios.decorators import allowed_methods
-from typing_extensions import Doc, Annotated  # type: ignore
-from nexios.structs import URLPath, RouteParam
-from nexios.http import Request, Response
-from nexios.http.response import JSONResponse
-from nexios.types import Scope, Send, Receive, ASGIApp
-from .converters import Convertor, CONVERTOR_TYPES, get_route_path
-from nexios.websockets import WebSocket
-from nexios.exceptions import NotFoundException
-from nexios.websockets.errors import WebSocketErrorMiddleware
+import warnings
+import typing
+import re
+from dataclasses import dataclass
 from pydantic import BaseModel
-from nexios.http.response import BaseResponse
-from nexios._internals._response_transformer import request_response
-from nexios._internals._route_builder import RouteType, RouteBuilder
-from nexios._internals.__middleware import (
-    wrap_middleware,
-    ASGIRequestResponseBridge,
-    DefineMiddleware as Middleware,
-)
-
+from .base import BaseRouter,BaseRoute
+from nexios.http import Request, Response
+from nexios.types import Scope, Receive, Send, ASGIApp, MiddlewareType, HandlerType
+from nexios.exceptions import NotFoundException
+from nexios.http.response import JSONResponse
 from nexios.dependencies import inject_dependencies
+from nexios._internals._route_builder import RouteBuilder
+from nexios._internals.__middleware import wrap_middleware,DefineMiddleware as Middleware,ASGIRequestResponseBridge
+from nexios.openapi import Parameter
+from nexios.structs import URLPath,RouteParam
+from nexios._internals._response_transformer import request_response
+from nexios.events import AsyncEventEmitter
+from nexios.decorators import allowed_methods
+from ._utils import get_route_path
 
-
-T = TypeVar("T")
 allowed_methods_default = ["get", "post", "delete", "put", "patch", "options"]
 
-
-def websocket_session(
-    func: typing.Callable[[WebSocket], typing.Awaitable[None]],
-) -> ASGIApp:
-    """
-    Takes a coroutine `func(session)`, and returns an ASGI application.
-    """
-    assert asyncio.iscoroutinefunction(func), "WebSocket endpoints must be async"
-
-    async def app(scope: Scope, receive: Receive, send: Send) -> None:
-        session = WebSocket(scope, receive=receive, send=send)
-
-        async def app(scope: Scope, receive: Receive, send: Send) -> None:
-            await func(session)
-
-        # await wrap_app_handling_exceptions(app, session)(scope, receive, send)
-        await app(scope, receive, send)
-
-    return app
-
-
-class BaseRouter(ABC):
-    """
-    Base class for routers. This class should not be instantiated directly.
-    Subclasses should implement the `__call__` method to handle specific routing logic.
-    """
-
-    def __init__(self, prefix: Optional[str] = None):
-        self.prefix = prefix or ""
-        self.routes: List[Any] = []
-        self.middlewares: List[Any] = []
-        self.sub_routers: Dict[str, ASGIApp] = {}
-
-        if self.prefix and not self.prefix.startswith("/"):
-            warnings.warn("Router prefix should start with '/'")
-            self.prefix = f"/{self.prefix}"
-
-    @abstractmethod
-    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
-        """
-        Abstract method to handle incoming requests. Subclasses must implement this method.
-
-        Args:
-            scope: The ASGI scope dictionary.
-            receive: The ASGI receive callable.
-            send: The ASGI send callable.
-
-        Raises:
-            NotImplementedError: If the method is not implemented by a subclass.
-        """
-        raise NotImplementedError("Subclasses must implement this method")
-
-    def add_middleware(self, middleware: MiddlewareType) -> None:
-        """
-        Add middleware to the router.
-
-        Args:
-            middleware: The middleware to add.
-        """
-
-        self.middlewares.append(middleware)
-
-    def build_middleware_stack(self, app: ASGIApp) -> ASGIApp:
-        """
-        Builds the middleware stack by applying all registered middlewares to the app.
-
-        Args:
-            app: The base ASGI application.
-
-        Returns:
-            ASGIApp: The application wrapped with all middlewares.
-        """
-        for mdw in reversed(self.middlewares):
-            app = mdw(app)
-        return app
-
-    def mount_router(
-        self, app: Union["Router", "WSRouter"], path: Optional[str] = None
-    ) -> None:
-        """
-        Mount an ASGI application (e.g., another Router) under a specific path prefix.
-
-        Args:
-            path: The path prefix under which the app will be mounted.
-            app: The ASGI application (e.g., another Router) to mount.
-        """
-        if not path:
-            path = app.prefix
-        path = path.rstrip("/")
-
-        if path == "":
-            self.sub_routers[path] = app
-            return
-        if not path.startswith("/"):
-            path = f"/{path}"
-
-        self.sub_routers[path] = app
-
-    def __repr__(self) -> str:
-        return f"<BaseRouter prefix='{self.prefix}' routes={len(self.routes)}>"
-
-
-class Routes:
+class Routes(BaseRoute):
     """
     Encapsulates all routing information for an API endpoint, including path handling,
     validation, OpenAPI documentation, and request processing.
@@ -381,15 +256,7 @@ class Routes:
 
         await app(scope, receive, send)
 
-    def __call__(self) -> Tuple[Pattern[str], HandlerType]:
-        """
-        Return the route components for registration.
-
-        Returns:
-            Tuple[Pattern[str], HandlerType]: The compiled regex pattern and the handler.
-
-        """
-        return self.pattern, self.handler
+       
 
     def __repr__(self) -> str:
         """
@@ -401,20 +268,24 @@ class Routes:
         return f"<Route {self.raw_path} methods={self.methods}>"
 
 
+
+
+
+
 class Router(BaseRouter):
     def __init__(
         self,
         prefix: Optional[str] = None,
-        routes: Optional[List[Routes]] = None,
+        routes: Optional[List[Union[Routes, type[BaseRoute]]]] = None,
         tags: Optional[List[str]] = None,
         exclude_from_schema: bool = False,
         name: Optional[str] = None,
     ):
         self.prefix = prefix or ""
         self.prefix.rstrip("/")
-        self.routes: List[Routes] = list(routes) if routes else []
+        self.routes = list(routes) if routes else []
         self.middlewares: typing.List[Middleware] = []
-        self.sub_routers: Dict[str, Router] = {}
+        self.sub_routers: Dict[str, Union[Router,ASGIApp]] = {}
         self.route_class = Routes
         self.tags = tags or []
         self.exclude_from_schema = exclude_from_schema
@@ -462,6 +333,9 @@ class Router(BaseRouter):
             app.add_route(route)
             ```
         """
+
+        if not isinstance(route, Routes):
+            raise TypeError("Route must be an instance of Routes")
         route.tags = self.tags + route.tags if route.tags else self.tags
         # original_handler = route.handler
 
@@ -2152,6 +2026,7 @@ class Router(BaseRouter):
             if url.startswith(mount_path):
                 scope["path"] = url[len(mount_path) :]
                 await sub_app(scope, receive, send)
+                
                 return
 
         path_matched = False
@@ -2225,233 +2100,18 @@ class Router(BaseRouter):
 
         return all_routes
 
-
-class WebsocketRoutes:
-    def __init__(
+    def register(
         self,
-        path: str,
-        handler: WsHandlerType,
-        middlewares: typing.List[WsMiddlewareType] = [],
+        app :ASGIApp,
+        prefix: str = "",
     ):
-        assert callable(handler), "Route handler must be callable"
-        assert asyncio.iscoroutinefunction(handler), "Route handler must be async"
-        self.raw_path = path
-        self.handler: WsHandlerType = inject_dependencies(handler)
-        self.middlewares = middlewares
-        self.route_info = RouteBuilder.create_pattern(path)
-        self.pattern = self.route_info.pattern
-        self.param_names = self.route_info.param_names
-        self.route_type = self.route_info.route_type
-        self.router_middleware = None
-
-    def match(self, path: str) -> typing.Tuple[Any, Any]:
         """
-        Match a path against this route's pattern and return captured parameters.
+        Register an ASGI application (e.g., another Router) under a specific path prefix.
 
         Args:
-            path: The URL path to match.
-
-        Returns:
-            Optional[Dict[str, Any]]: A dictionary of captured parameters if the path matches,
-            otherwise None.
+            app: The ASGI application (e.g., another Router) to register.
+            prefix: The path prefix under which the app will be registered.
         """
-        match = self.pattern.match(path)
-        if match:
-            matched_params = match.groupdict()
-            for key, value in matched_params.items():
-                matched_params[key] = self.route_info.convertor[  # type:ignore
-                    key
-                ].convert(value)
-            return match, matched_params
-        return None, None
-
-    async def handle(self, scope: Scope, receive: Receive, send: Send) -> None:
-        """
-        Handles the WebSocket connection by calling the route's handler.
-
-        Args:
-            websocket: The WebSocket connection.
-            params: The extracted route parameters.
-        """
-        websocket_session = WebSocket(scope, receive=receive, send=send)
-        await self.handler(websocket_session)
-
-    def __repr__(self) -> str:
-        return f"<WSRoute {self.raw_path}>"
-
-
-class WSRouter(BaseRouter):
-    def __init__(
-        self,
-        prefix: Optional[str] = None,
-        middleware: Optional[List[Any]] = [],
-        routes: Optional[List[WebsocketRoutes]] = [],
-    ):
-        self.prefix = prefix or ""
-        self.routes: List[WebsocketRoutes] = routes or []
-        self.middlewares: List[Callable[[ASGIApp], ASGIApp]] = []
-        self.sub_routers: Dict[str, ASGIApp] = {}
-        if self.prefix and not self.prefix.startswith("/"):
-            warnings.warn("WSRouter prefix should start with '/'")
-            self.prefix = f"/{self.prefix}"
-
-    def add_ws_route(
-        self,
-        route: Annotated[
-            WebsocketRoutes,
-            Doc("An instance of the Routes class representing a WebSocket route."),
-        ],
-    ) -> None:
-        """
-        Adds a WebSocket route to the application.
-
-        This method registers a WebSocket route, allowing the application to handle WebSocket connections.
-
-        Args:
-            route (Routes): The WebSocket route configuration.
-
-        Returns:
-            None
-
-        Example:
-            ```python
-            route = Routes("/ws/chat", chat_handler)
-            app.add_ws_route(route)
-            ```
-        """
-        self.routes.append(route)
-
-    def add_ws_middleware(self, middleware: type[ASGIApp]) -> None:  # type: ignore[override]
-        """Add middleware to the WebSocket router"""
-        self.middlewares.insert(0, middleware)  # type: ignore
-
-    def ws_route(
-        self,
-        path: Annotated[
-            str, Doc("The WebSocket route path. Must be a valid URL pattern.")
-        ],
-        handler: Annotated[
-            Optional[WsHandlerType],
-            Doc("The WebSocket handler function. Must be an async function."),
-        ] = None,
-        middlewares: Annotated[
-            List[WsMiddlewareType],
-            Doc("List of middleware to be executes before the router handler"),
-        ] = [],
-    ) -> Any:
-        """
-        Registers a WebSocket route.
-
-        This decorator is used to define WebSocket routes in the application, allowing handlers
-        to manage WebSocket connections. When a WebSocket client connects to the given path,
-        the specified handler function will be executed.
-
-        Returns:
-            Callable: The original WebSocket handler function.
-
-        Example:
-            ```python
-
-            @app.ws_route("/ws/chat")
-            async def chat_handler(websocket):
-                await websocket.accept()
-                while True:
-                    message = await websocket.receive_text()
-                    await websocket.send_text(f"Echo: {message}")
-            ```
-        """
-        if handler:
-            return self.add_ws_route(
-                WebsocketRoutes(path, handler, middlewares=middlewares)
-            )
-
-        def decorator(handler: WsHandlerType) -> WsHandlerType:
-            self.add_ws_route(WebsocketRoutes(path, handler, middlewares=middlewares))
-            return handler
-
-        return decorator
-
-    def build_middleware_stack(  # type:ignore
-        self, scope: Scope, receive: Receive, send: Send
-    ) -> ASGIApp:  # type:ignore
-        app = self.app
-        for mdw in reversed(self.middlewares):
-            app = mdw(app)  # type:ignore[assignment]
-        return app
-
-    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
-        if scope["type"] != "websocket":
-            return
-        app = self.build_middleware_stack(scope, receive, send)
-        app = WebSocketErrorMiddleware(app)
-        await app(scope, receive, send)
-
-    async def app(self, scope: Scope, receive: Receive, send: Send) -> None:
-
-        url = get_route_path(scope)
-        for mount_path, sub_app in self.sub_routers.items():
-            if url.startswith(mount_path):
-                scope["path"] = url[len(mount_path) :]
-                await sub_app(scope, receive, send)
-                return
-        for route in self.routes:
-            match, params = route.match(url)
-            if match:
-
-                scope["route_params"] = params
-                await route.handle(scope, receive, send)
-                return
-        await send({"type": "websocket.close", "code": 404})
-
-    def wrap_asgi(
-        self,
-        middleware_cls: Annotated[
-            Callable[[ASGIApp], Any],
-            Doc(
-                "An ASGI middleware class or callable that takes an app as its first argument and returns an ASGI app"
-            ),
-        ],
-    ) -> None:
-        """
-        Wraps the entire application with an ASGI middleware.
-
-        This method allows adding middleware at the ASGI level, which intercepts all requests
-        (HTTP, WebSocket, and Lifespan) before they reach the application.
-
-        Args:
-            middleware_cls: An ASGI middleware class or callable that follows the ASGI interface
-            *args: Additional positional arguments to pass to the middleware
-            **kwargs: Additional keyword arguments to pass to the middleware
-
-        Returns:
-            NexiosApp: The application instance for method chaining
-
-
-        """
-        self.app = middleware_cls(self.app)
-
-    def mount_router(  # type:ignore
-        self, app: "WSRouter", path: typing.Optional[str] = None
-    ) -> None:  # type:ignore
-        """
-        Mount an ASGI application (e.g., another Router) under a specific path prefix.
-
-        Args:
-            path: The path prefix under which the app will be mounted.
-            app: The ASGI application (e.g., another Router) to mount.
-        """
-
-        if not path:
-            path = app.prefix
-        path = path.rstrip("/")
-
-        if path == "":
-            self.sub_routers[path] = app
-            return
-        if not path.startswith("/"):
-            path = f"/{path}"
-
-        self.sub_routers[path] = app
-
-    def __repr__(self) -> str:
-        return f"<WSRouter prefix='{self.prefix}' routes={len(self.routes)}>"
+        
+        self.sub_routers[prefix] = app
+
